Transfer Request API

+++ src/main/java/com/samadhan/enums/BikeModelEnum.java
@@ -3,19 +3,19 @@ package com.samadhan.enums;
 
 public enum BikeModelEnum {
 
-    Commuter(0, "Commuter", 120),
-    Sports(1, "Sports", 180),
-    Cruiser(2, "Cruiser", 220),
-    Touring(3, "Touring", 280),
-    Adventure(4, "Adventure", 210),
-    Scooter(5, "Scooter", 105),
-    Electric(6, "Electric", 140);
+    Commuter(0, "Commuter", 120.0),
+    Sports(1, "Sports", 180.0),
+    Cruiser(2, "Cruiser", 220.0),
+    Touring(3, "Touring", 280.0),
+    Adventure(4, "Adventure", 210.0),
+    Scooter(5, "Scooter", 105.0),
+    Electric(6, "Electric", 140.0);
 
     private final int id;
     private final String label;
-    private final int averageWeightKg;
+    private final Double averageWeightKg;
 
-    BikeModelEnum(int id, String label, int averageWeightKg) {
+    BikeModelEnum(int id, String label, Double averageWeightKg) {
         this.id = id;
         this.label = label;
         this.averageWeightKg = averageWeightKg;
@@ -29,7 +29,7 @@ public enum BikeModelEnum {
         return label;
     }
 
-    public int getAverageWeightKg() {
+    public Double getAverageWeightKg() {
         return averageWeightKg;
     }
